refactor: 학과공지 크롤링 및 스케줄링 코드 리팩터링
- 디렉터리 및 파일명 유지보수를 위한 구체적으로 지정
